backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pillow
torch
transformers
google-generativeai
pyttsx3
opencv-python
python-multipart
